ES6ify ClientManager
Just change to ES6 class and rename filename.
