+++ src/command-line/reset.js
@@ -1,5 +1,5 @@
 import bcrypt from 'bcrypt-nodejs';
-import ClientManager from '../clientManager';
+import ClientManager from '../ClientManager';
 import fs from 'fs';
 import program from 'commander';
 import Helper from '../helper';
